requirements.txt
fastapi
uvicorn
python-dotenv
httpx
google-generativeai
jinja2
loguru
